+++ test/src/test/java/de/njsm/tinyauth/test/ApiTests.java
@@ -17,13 +17,25 @@
 
 package de.njsm.tinyauth.test;
 
+import com.nimbusds.jose.JWSAlgorithm;
+import com.nimbusds.jose.JWSHeader;
+import com.nimbusds.jose.JWSSigner;
+import com.nimbusds.jose.crypto.ECDSASigner;
+import com.nimbusds.jose.crypto.MACSigner;
+import com.nimbusds.jose.jwk.JWK;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
 import de.njsm.tinyauth.test.data.Client;
 import de.njsm.tinyauth.test.data.User;
 import de.njsm.tinyauth.test.repository.Clients;
+import de.njsm.tinyauth.test.repository.Endpoints;
 import de.njsm.tinyauth.test.repository.Users;
 import io.restassured.path.json.JsonPath;
+import io.restassured.response.ValidatableResponse;
 import org.junit.jupiter.api.Test;
 
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
 import java.util.Set;
 
 import static de.njsm.tinyauth.test.oidc.Identifiers.*;
@@ -33,21 +45,29 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ApiTests implements TinyAuthTest {
 
-    private final Client client = Clients.getConfidentialClient();
+    private Client client = Clients.getConfidentialClient();
 
     private final User user = Users.getUser();
 
-    private final Set<String> scopes = Set.of("openid");
+    private Set<String> scopes = Set.of("openid");
 
     @Test
     void testClientCredentialsGrant() throws Exception {
+        authenticate(client, tokenEndpoint().requestWithClientCredentials(client, scopes));
+    }
+
+    @Test
+    void clientCredentialsWithDisallowedScopeIsFiltered() throws Exception {
+        scopes = Set.of("openid", "email");
+        client = Clients.getRestrictedScopeClient();
+
         JsonPath tokenResponse = tokenEndpoint().requestWithClientCredentials(client, scopes)
                 .statusCode(200)
                 .body(EXPIRES_IN, equalTo(60))
                 .body(TOKEN_TYPE, equalToIgnoringCase(TOKEN_TYPE_CONTENT))
                 .extract().body().jsonPath();
 
-        assertEquals(scopes, Set.of(tokenResponse.getString(SCOPE).split(" ")));
+        assertEquals(Set.of("openid"), Set.of(tokenResponse.getString(SCOPE).split(" ")));
         assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
         tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, client);
     }
@@ -61,6 +81,22 @@ public class ApiTests implements TinyAuthTest {
                 .extract().body().jsonPath();
 
         assertEquals(scopes, Set.of(tokenResponse.getString(SCOPE).split(" ")));
+        assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
+        tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, user);
+    }
+
+    @Test
+    void passwordGrantsDisallowedScopeIsFiltered() throws Exception {
+        scopes = Set.of("openid", "email");
+        client = Clients.getRestrictedScopeClient();
+
+        JsonPath tokenResponse = tokenEndpoint().requestWithPassword(client, user, scopes)
+                .statusCode(200)
+                .body(EXPIRES_IN, equalTo(60))
+                .body(TOKEN_TYPE, equalToIgnoringCase(TOKEN_TYPE_CONTENT))
+                .extract().body().jsonPath();
+
+        assertEquals(Set.of("openid"), Set.of(tokenResponse.getString(SCOPE).split(" ")));
         assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
         tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, user);
     }
@@ -83,5 +119,64 @@ public class ApiTests implements TinyAuthTest {
                 .statusCode(400)
                 .body(ERROR, equalTo(INVALID_GRANT))
                 .body(ERROR_DESCRIPTION, equalTo("rate limited"));
+    }
+
+    @Test
+    void clientCredentialsWorkWithSharedSecret() throws Exception {
+        Client client = Clients.getAdvancedAuthClient();
+        String token = buildTokenFromSharedSecret(client);
+        authenticate(client, tokenEndpoint().requestWithClientCredentialsToken(token, scopes));
+    }
+
+    @Test
+    void clientCredentialsWorkWithPublicKey() throws Exception {
+        Client client = Clients.getAdvancedAuthClient();
+        String token = buildTokenFromPrivateKey(client);
+        authenticate(client, tokenEndpoint().requestWithClientCredentialsToken(token, scopes));
+    }
+
+    private void authenticate(Client client, ValidatableResponse validatableResponse) throws Exception {
+        JsonPath tokenResponse = validatableResponse
+                .statusCode(200)
+                .body(EXPIRES_IN, equalTo(60))
+                .body(TOKEN_TYPE, equalToIgnoringCase(TOKEN_TYPE_CONTENT))
+                .extract().body().jsonPath();
+
+        assertEquals(scopes, Set.of(tokenResponse.getString(SCOPE).split(" ")));
+        assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
+        tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, client);
+    }
+
+    private String buildTokenFromSharedSecret(Client client) throws Exception {
+        Date now = new Date();
+        JWSSigner signer = new MACSigner(client.getPassword().getBytes(StandardCharsets.UTF_8));
+        JWTClaimsSet claimsSet = new JWTClaimsSet.Builder()
+                .subject(client.getClientId())
+                .issuer(client.getClientId())
+                .audience(Endpoints.getTokenUrl())
+                .jwtID("jti")
+                .expirationTime(new Date(now.getTime() + 5 * 1000))
+                .issueTime(now)
+                .build();
+        SignedJWT signedJWT = new SignedJWT(new JWSHeader(JWSAlgorithm.HS256), claimsSet);
+        signedJWT.sign(signer);
+        return signedJWT.serialize();
+    }
+
+    private String buildTokenFromPrivateKey(Client client) throws Exception {
+        Date now = new Date();
+        JWK jwk = JWK.parseFromPEMEncodedObjects(client.getPublicKey());
+        JWSSigner signer = new ECDSASigner(jwk.toECKey());
+        JWTClaimsSet claimsSet = new JWTClaimsSet.Builder()
+                .subject(client.getClientId())
+                .issuer(client.getClientId())
+                .audience(Endpoints.getTokenUrl())
+                .jwtID("jti")
+                .expirationTime(new Date(now.getTime() + 5 * 1000))
+                .issueTime(now)
+                .build();
+        SignedJWT signedJWT = new SignedJWT(new JWSHeader(JWSAlgorithm.ES384), claimsSet);
+        signedJWT.sign(signer);
+        return signedJWT.serialize();
     }
 }