+++ test/src/test/java/de/njsm/tinyauth/test/repository/Clients.java
@@ -32,7 +32,27 @@ public class Clients {
         return new Client("needed-for-token-switch-attack", "password", "http://localhost:34345/redirect/needed-for-token-switch-attack");
     }
 
+    public static Client getAdvancedAuthClient() {
+        return new Client("advanced-client-auth", "passwordpasswordpasswordpasswordpasswordpassword",
+                "http://localhost:34345/redirect/advanced-client-auth",
+                "-----BEGIN PRIVATE KEY-----\n" +
+                "MIG2AgEAMBAGByqGSM49AgEGBSuBBAAiBIGeMIGbAgEBBDDcQ4w1qyBpSr7kHk7M\n" +
+                "lK6tEOBDE0gTozttlqB6qW71JcfhGp7YxECP13XUa/XfkNuhZANiAARfBFqROixP\n" +
+                "tsrlgQ5GVCMvOkd6GS4sMS7/SkXZrFsfIAZq0PZAJ/Qp+a7KUGV2jJqlKPyYJjXU\n" +
+                "Se+nShnhLWRiZuF+AgTBfXq4OQqsyxpN4j+22+BHTXHMx89fKMDgh+w=\n" +
+                "-----END PRIVATE KEY-----\n" +
+                "-----BEGIN PUBLIC KEY-----\n" +
+                "MHYwEAYHKoZIzj0CAQYFK4EEACIDYgAEXwRakTosT7bK5YEORlQjLzpHehkuLDEu\n" +
+                "/0pF2axbHyAGatD2QCf0KfmuylBldoyapSj8mCY11Envp0oZ4S1kYmbhfgIEwX16\n" +
+                "uDkKrMsaTeI/ttvgR01xzMfPXyjA4Ifs\n" +
+                "-----END PUBLIC KEY-----\n");
+    }
+
     public static Client getPublicClient() {
         return new Client("public", "", "http://localhost:34345/redirect/public");
+    }
+
+    public static Client getRestrictedScopeClient() {
+        return new Client("restricted-scopes", "password", "http://localhost:34345/redirect/restricted-scopes");
     }
 }